+++ libs/codec/lpc10/encode.c
@@ -279,10 +279,10 @@ static integer c__2 = 2;
 
     /* Function Body */
 /*  Scale RMS and RC's to integers */
-    *irms = *rms;
+    *irms = (integer)*rms;
     i__1 = contrl_1.order;
     for (i__ = 1; i__ <= i__1; ++i__) {
-	irc[i__] = rc[i__] * 32768.f;
+	irc[i__] = (integer)(rc[i__] * 32768.f);
     }
 /* 	IF(LISTL.GE.3)WRITE(FDEBUG,800)VOICE,PITCH,IRMS,(IRC(I),I=1,ORDER) */
 /* 800	FORMAT(1X,/,' <<ENCODE IN>>',T32,2I3,I6,I5,T50,10I8) */
@@ -336,8 +336,8 @@ static integer c__2 = 2;
     i__1 = contrl_1.order;
     for (i__ = 3; i__ <= i__1; ++i__) {
 	i2 = irc[i__] / 2;
-	i2 = (i2 + enadd[contrl_1.order + 1 - i__ - 1]) * enscl[
-		contrl_1.order + 1 - i__ - 1];
+	i2 = (integer)((i2 + enadd[contrl_1.order + 1 - i__ - 1]) * enscl[
+		contrl_1.order + 1 - i__ - 1]);
 /* Computing MIN */
 	i__2 = max(i2,-127);
 	i2 = min(i__2,127);
