+++ libs/codec/lpc10/tbdm.c
@@ -121,7 +121,7 @@ extern int tbdm_(real *speech, integer *lpita, integer *tau, integer *ltau, real
     difmag_(&speech[1], lpita, &tau[1], ltau, &tau[*ltau], &amdf[1], minptr, 
 	    maxptr);
     *mintau = tau[*minptr];
-    minamd = amdf[*minptr];
+    minamd = (integer)amdf[*minptr];
 /*   Build table containing all lags within +/- 3 of the AMDF minimum */
 /*    excluding all that have already been computed */
     ltau2 = 0;
@@ -147,7 +147,7 @@ extern int tbdm_(real *speech, integer *lpita, integer *tau, integer *ltau, real
 		maxp2);
 	if (amdf2[minp2 - 1] < (real) minamd) {
 	    *mintau = tau2[minp2 - 1];
-	    minamd = amdf2[minp2 - 1];
+	    minamd = (integer)amdf2[minp2 - 1];
 	}
     }
 /*   Check one octave up, if there are any lags not yet computed */
@@ -165,7 +165,7 @@ extern int tbdm_(real *speech, integer *lpita, integer *tau, integer *ltau, real
 		maxp2);
 	if (amdf2[minp2 - 1] < (real) minamd) {
 	    *mintau = tau2[minp2 - 1];
-	    minamd = amdf2[minp2 - 1];
+	    minamd = (integer)amdf2[minp2 - 1];
 	    *minptr += -20;
 	}
     }
